### 2018-05-20 #744 @whedon
[PRE REVIEW]: opt_einsum: A tensor contraction order optimizer for NumPy
**Submitting author:** @dgasmith (<a href="http://orcid.org/0000-0001-8626-0900">Daniel Smith</a>)
**Repository:** <a href="https://github.com/dgasmith/opt_einsum" target ="_blank">https://github.com/dgasmith/opt_einsum</a>
**Version:** v2.0.0
