### 2018-05-20 #744 @arfon
@dgasmith - it looks like you've got unmatched quotations on the first line of your paper:

```
title: 'opt_einsum: A Python package for optimizing contraction order for einsum-like expressions`

```
