### 2018-05-20 7c93c57
Fix small paper typo
