### 2018-05-21 #744 @dgasmith
There we are, seems it dislikes syntax highlighting or semicolons in the title.
