### 2018-05-25 #753 @pjotrp
@dgasmith thanks for your submission! To expedite review, do you mind going through above check list to make sure we can tick all boxes? You can't tick them, just go through them.
