### 2018-05-25 05e681e
Create CONTRIBUTING.md
