### 2018-05-26 #753 @dgasmith
@pjotrp Thanks, I believe I have them all except `Community guidelines` might be weak. Do you have examples of good repositories that do this? Would be interesting to do some sort of guidelines with [URSSI](http://urssi.us) or similar organizations.
