### 2018-05-26 68c6ac4
Expands upon the contrib guidelines
