### 2018-05-27 #753 @pjotrp
Great @dgasmith. You can lift them from most papers in JOSS. It is not exactly a requirement, but a good to have. @prasunanand, we are ready for review!
