### 2018-06-14 b89a65f
Test: Breaks edge case tests out of test_contract
