### 2018-06-18 #753 @prasunanand
Sorry for reviewing late. I am reviewing it today.

@dgasmith : I would suggest to add the link to './github/Contributing.md' in '/ReadMe.md' .

You also need to edit the [Contributing.md](https://github.com/dgasmith/opt_einsum/blob/master/.github/CONTRIBUTING.md). For example, in 1st paragraph `project_name` should be replaced with `opt_einsum` .
