### 2018-06-18 eda3a4f
JOSS: Minor documentation tweaks
