### 2018-06-18 05de336
fix perf regression due to addition of many symbols
