### 2018-06-18 26eecf9
allow ~ 7500 unique indices
