+++ opt_einsum/parser.py
@@ -2,20 +2,10 @@
 A functionally equivalent parser of the numpy.einsum input parser
 """
 
-import sys
-
 import numpy as np
 
-einsum_symbols_base = 'abcdefghijklmnopqrstuvwxyz'
-einsum_symbols = einsum_symbols_base + 'ABCDEFGHIJKLMNOPQRSTUVWXYZ'
 
-# boost the number of symbols using unicode if python3
-if sys.version_info[0] >= 3:
-    einsum_symbols += ''.join(map(chr, range(193, 688)))
-    einsum_symbols += ''.join(map(chr, range(913, 1367)))
-    einsum_symbols += ''.join(map(chr, range(13312, 19893)))
-
-einsum_symbols_set = set(einsum_symbols)
+einsum_symbols_base = 'abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ'
 
 
 def is_valid_einsum_char(x):
@@ -30,6 +20,28 @@ def has_valid_einsum_chars_only(einsum_str):
     return all(map(is_valid_einsum_char, einsum_str))
 
 
+def symbol(i):
+    """Get the symbol corresponding to int ``i`` - runs through the usual 52
+    letters before resorting to unicode characters, starting at ``chr(192)``.
+    """
+    if i < 52:
+        return einsum_symbols_base[i]
+    return chr(i + 140)
+
+
+def gen_unused_symbols(used, n):
+    """Generate ``n`` symbols that are not already in ``used``.
+    """
+    i = cnt = 0
+    while cnt < n:
+        s = symbol(i)
+        i += 1
+        if s in used:
+            continue
+        yield s
+        cnt += 1
+
+
 def convert_to_valid_einsum_chars(einsum_str):
     """Convert the str ``einsum_str`` to contain only the alphabetic characters
     valid for numpy einsum.
@@ -40,7 +52,7 @@ def convert_to_valid_einsum_chars(einsum_str):
         (valid if is_valid_einsum_char(x) else invalid).add(x)
 
     # get replacements for invalid chars that are not already used
-    available = (x for x in einsum_symbols if x not in valid)
+    available = gen_unused_symbols(valid, len(invalid))
 
     # map invalid to available and replace in the inputs
     replacer = dict(zip(invalid, available))
@@ -53,8 +65,6 @@ def find_output_str(subscripts):
     tmp_subscripts = subscripts.replace(",", "")
     output_subscript = ""
     for s in sorted(set(tmp_subscripts)):
-        if s not in einsum_symbols_set:
-            raise ValueError("Character %s is not a valid symbol." % s)
         if tmp_subscripts.count(s) == 1:
             output_subscript += s
     return output_subscript
@@ -67,19 +77,6 @@ def possibly_convert_to_numpy(x):
         return np.asanyarray(x)
     else:
         return x
-
-
-def gen_unused_symbols(used, n):
-    """Generate ``n`` symbols that are not already in ``used``:
-    """
-    cnt = 0
-    for s in einsum_symbols:
-        if s in used:
-            continue
-        yield s
-        cnt += 1
-        if cnt == n:
-            break
 
 
 def parse_einsum_input(operands):
@@ -115,13 +112,6 @@ def parse_einsum_input(operands):
         subscripts = operands[0].replace(" ", "")
         operands = [possibly_convert_to_numpy(x) for x in operands[1:]]
 
-        # Ensure all characters are valid
-        for s in subscripts:
-            if s in '.,->':
-                continue
-            if s not in einsum_symbols_set:
-                raise ValueError("Character %s is not a valid symbol." % s)
-
     else:
         tmp_operands = list(operands)
         operand_list = []
[…]
