### 2018-06-19 b0d9c86
rename symbol -> get_symbol and document it
