### 2018-06-19 1fe9b87
specify utf8 encoding for sake of python2
