+++ opt_einsum/parser.py
@@ -1,3 +1,5 @@
+#!/usr/bin/env python
+# coding: utf-8
 """
 A functionally equivalent parser of the numpy.einsum input parser
 """
