### 2018-06-25 #753 @prasunanand
@pjotrp : Review complete :)

@dgasmith : Cool software. Nice work :muscle:

```
timeit contract('pi,qj,ijkl,rk,sl->pqrs', C, C, I, C, C)
The slowest run took 393.49 times longer than the fastest. This could mean that an intermediate result is being cached.
1000 loops, best of 3: 411 µs per loop
```
I am not sure if the message "regarding caching" could be suppressed.
