### 2018-06-26 #753 @pjotrp
@dgasmith can you fix the PDF generation? Just rerun using above whedon command.
