### 2018-06-26 #753 @dgasmith
@prasunanand I could not find this in our documentation, did we miss it? As a note `%timeit` is a Jupyter magic function that we do not have control over.

Thanks for the review and glad you like the work!

----

@pjotrp Could you give me a hint on what might be going wrong? This error looks a bit more vague than the previous failures and the article does not have a `$` in it.
