### 2018-06-26 #753 @arfon
> @pjotrp Could you give me a hint on what might be going wrong? This error looks a bit more vague than the previous failures and the article does not have a $ in it.

It seems like the underscore in the title is causing issues with Pandoc. I've fixed the issue in https://github.com/dgasmith/opt_einsum/pull/28
