### 2018-06-26 #753 @prasunanand
@dgasmith, You didn't miss anything. I was just wondering if it was possible to control `%timeit` to suppress the "caching" message.
