### 2018-06-26 #753 @dgasmith
@prasunanand Unfortunately I do not see it in their [options](https://ipython.readthedocs.io/en/stable/interactive/magics.html#magic-timeit).
