### 2018-06-28 #753 @dgasmith
@arfon Great: https://zenodo.org/record/1299981

As a note I made a minor typo fix to the paper here: https://github.com/dgasmith/opt_einsum/commit/b08d789873852074833b608b3ba665b2816d4eae
