### 2018-07-10 cdf7a78
allow constants in expressions
