### 2018-07-11 81ce5ee
enable constants for backends
