### 2018-07-13 01ea129
add constants support for tensorflow and theano
theano use tensor.constant
