+++ opt_einsum/tests/test_backends.py
@@ -6,6 +6,7 @@ from opt_einsum import contract, helpers, contract_expression, backends
 try:
     import tensorflow as tf
     found_tensorflow = True
+    sess = tf.Session()
 except ImportError:
     found_tensorflow = False
 
@@ -57,15 +58,85 @@ def test_tensorflow(string):
     shps = [v.shape for v in views]
     expr = contract_expression(string, *shps, optimize=True)
 
-    sess = tf.Session()
     with sess.as_default():
         expr(*views, backend='tensorflow', out=opt)
 
     assert np.allclose(ein, opt)
 
     # test non-conversion mode
-    tensorflow_views = backends.convert_arrays_to_tensorflow(views)
+    tensorflow_views = [backends.to_tensorflow(view) for view in views]
     expr(*tensorflow_views, backend='tensorflow')
+
+
+@pytest.mark.skipif(not found_tensorflow, reason="Tensorflow not installed.")
+def test_tensorflow_with_constants():
+    eq = 'ij,jk,kl->li'
+    shapes = (2, 3), (3, 4), (4, 5)
+    constants = {0, 2}
+    ops = [np.random.rand(*shp) if i in constants else shp for i, shp in enumerate(shapes)]
+    var = np.random.rand(*shapes[1])
+
+    res_exp = contract(eq, ops[0], var, ops[2])
+
+    expr = contract_expression(eq, *ops, constants=constants)
+
+    # check tensorflow
+    with sess.as_default():
+        res_got = expr(var, backend='tensorflow')
+    assert 'tensorflow' in expr._parsed_constants
+    assert np.allclose(res_exp, res_got)
+
+    # check can call with numpy still
+    res_got2 = expr(var, backend='numpy')
+    assert np.allclose(res_exp, res_got2)
+
+    # check tensorflow call returns tensorflow still
+    res_got3 = expr(backends.to_tensorflow(var), backend='tensorflow')
+    assert isinstance(res_got3, tf.Tensor)
+
+
+@pytest.mark.skipif(not found_theano, reason="Theano not installed.")
+@pytest.mark.parametrize("string", tests)
+def test_theano(string):
+    views = helpers.build_views(string)
+    ein = contract(string, *views, optimize=False, use_blas=False)
+    shps = [v.shape for v in views]
+
+    expr = contract_expression(string, *shps, optimize=True)
+
+    opt = expr(*views, backend='theano')
+    assert np.allclose(ein, opt)
+
+    # test non-conversion mode
+    theano_views = [backends.to_theano(view) for view in views]
+    theano_opt = expr(*theano_views, backend='theano')
+    assert isinstance(theano_opt, theano.tensor.TensorVariable)
+
+
+@pytest.mark.skipif(not found_theano, reason="theano not installed.")
+def test_theano_with_constants():
+    eq = 'ij,jk,kl->li'
+    shapes = (2, 3), (3, 4), (4, 5)
+    constants = {0, 2}
+    ops = [np.random.rand(*shp) if i in constants else shp for i, shp in enumerate(shapes)]
+    var = np.random.rand(*shapes[1])
+
+    res_exp = contract(eq, ops[0], var, ops[2])
+
+    expr = contract_expression(eq, *ops, constants=constants)
+
+    # check theano
+    res_got = expr(var, backend='theano')
+    assert 'theano' in expr._parsed_constants
+    assert np.allclose(res_exp, res_got)
+
+    # check can call with numpy still
+    res_got2 = expr(var, backend='numpy')
+    assert np.allclose(res_exp, res_got2)
+
+    # check theano call returns theano still
[…]
