+++ opt_einsum/backends.py
@@ -62,48 +62,97 @@ def has_einsum(backend):
 
 # Tensorflow
 
-def convert_arrays_to_tensorflow(arrays):
-    """Convert numpy arrays to ``tensorflow.placeholder`` instances.
+def to_tensorflow(array, constant=False):
+    """Convert a numpy array to a ``tensorflow.placeholder`` instance.
     """
     import tensorflow
-    return [tensorflow.placeholder(x.dtype, x.shape) for x in arrays]
+
+    if isinstance(array, numpy.ndarray):
+        if constant:
+            return tensorflow.constant(array, array.dtype, array.shape)
+
+        return tensorflow.placeholder(array.dtype, array.shape)
+
+    return array
 
 
 def build_tensorflow_expression(arrays, expr):
     """Build a tensorflow function based on ``arrays`` and ``expr``.
     """
     import tensorflow
-    placeholders = convert_arrays_to_tensorflow(arrays)
+
+    placeholders = [to_tensorflow(array) for array in arrays]
     graph = expr._normal_contract(placeholders, backend='tensorflow')
 
     def tensorflow_contract(*arrays):
         session = tensorflow.get_default_session()
-        return session.run(graph, feed_dict=dict(zip(placeholders, arrays)))
+        # only want to feed placeholders - constant tensors already have values
+        feed_dict = {p: a for p, a in zip(placeholders, arrays) if p.op.type == 'Placeholder'}
+        return session.run(graph, feed_dict=feed_dict)
 
     return tensorflow_contract
 
 
+def parse_constants_tensorflow(const_arrays, expr):
+    """Convert constant arguments to tensorflow constants, and perform any
+    possible constant contractions. Requires evaluating a tensorflow graph.
+    """
+    import tensorflow
+
+    # compute the partial graph of new inputs
+    const_arrays = [to_tensorflow(x, constant=True) for x in const_arrays]
+    new_ops, new_contraction_list = expr(*const_arrays, backend='tensorflow', parse_constants=True)
+
+    # evaluate the new inputs and convert to tensorflow constants
+    session = tensorflow.get_default_session()
+    new_ops = [None if x is None else to_tensorflow(session.run(x), constant=True) for x in new_ops]
+
+    return new_ops, new_contraction_list
+
+
 # Theano
 
-def convert_arrays_to_theano(arrays):
-    """Convert numpy arrays to ``theano.tensor.TensorType`` instances.
+def to_theano(array, constant=False):
+    """Convert a numpy array to ``theano.tensor.TensorType`` instance.
     """
     import theano
-    return [theano.tensor.TensorType(dtype=x.dtype, broadcastable=[False] * len(x.shape))() for x in arrays]
+
+    if isinstance(array, numpy.ndarray):
+        if constant:
+            return theano.tensor.constant(array)
+
+        return theano.tensor.TensorType(dtype=array.dtype, broadcastable=[False] * len(array.shape))()
+
+    return array
 
 
 def build_theano_expression(arrays, expr):
     """Build a theano function based on ``arrays`` and ``expr``.
     """
     import theano
-    in_vars = convert_arrays_to_theano(arrays)
+
+    in_vars = [to_theano(array) for array in arrays]
     out_var = expr._normal_contract(in_vars, backend='theano')
-    graph = theano.function(in_vars, out_var)
+
+    # don't supply constants to graph
+    graph_ins = [x for x in in_vars if not isinstance(x, theano.tensor.TensorConstant)]
+    graph = theano.function(graph_ins, out_var)
 
     def theano_contract(*arrays):
-        return graph(*arrays)
[…]
