### 2018-07-14 4d62ef0
doc updates
