+++ opt_einsum/contract.py
@@ -530,20 +530,25 @@ class ContractExpression:
     def __init__(self, contraction, contraction_list, constants_dict, **einsum_kwargs):
         self.contraction_list = contraction_list
         self.einsum_kwargs = einsum_kwargs
+        self.contraction = format_const_einsum_str(contraction, constants_dict.keys())
 
+        # need to know _full_num_args to parse constants with, and num_args to call with
         self._full_num_args = contraction.count(',') + 1
         self.num_args = self._full_num_args - len(constants_dict)
 
-        self.contraction = format_const_einsum_str(contraction, constants_dict.keys())
+        # likewise need to know full contraction list
+        self._full_contraction_list = contraction_list
 
         self._constants_dict = constants_dict
         self._parsed_constants = {}
-        self._full_contraction_list = contraction_list
+        self._backend_expressions = {}
 
     def parse_constants(self, backend):
         """Convert any constant operands to the correct backend form, and
         perform as many contractions as possible to create a new list of
-        operands, stored in ``self._parsed_constants[backend]``.
+        operands, stored in ``self._parsed_constants[backend]``. This also
+        makes sure ``self.contraction_list`` only contains the remaining,
+        non-const operations.
         """
         # prepare a list of operands, with `None` for non-consts
         tmp_const_ops = [self._constants_dict.get(i, None) for i in range(self._full_num_args)]
@@ -554,13 +559,25 @@ class ContractExpression:
         self.contraction_list = new_contraction_list
 
     def _get_parsed_constants(self, backend):
+        """Retrieve or generate the cached list of constant operators (mixed
+        in with None representing non-consts) and the remaining contraction
+        list.
+        """
         try:
             return self._parsed_constants[backend]
         except KeyError:
             self.parse_constants(backend)
             return self._parsed_constants[backend]
 
-    def _normal_contract(self, arrays, out=None, backend='numpy', parse_constants=False):
+    def _get_backend_expression(self, arrays, backend):
+        try:
+            return self._backend_expressions[backend]
+        except KeyError:
+            fn = backends.build_expression(backend, arrays, self)
+            self._backend_expressions[backend] = fn
+            return fn
+
+    def _contract(self, arrays, out=None, backend='numpy', parse_constants=False):
         """The normal, core contraction.
         """
         contraction_list = self._full_contraction_list if parse_constants else self.contraction_list
@@ -568,22 +585,20 @@ class ContractExpression:
         return _core_contract(list(arrays), contraction_list, out=out, backend=backend,
                               parse_constants=parse_constants, **self.einsum_kwargs)
 
-    def _convert_contract(self, arrays, out, backend, parse_constants=False):
+    def _contract_with_conversion(self, arrays, out, backend, parse_constants=False):
         """Special contraction, i.e. contraction with a different backend
-        but converting to and from that backend. Checks for
-        ``self._{backend}_contract``, generates it if is missing, then calls it
+        but converting to and from that backend. Retrieves or generates a
+        cached expression using ``arrays`` as templates, then calls it
         with ``arrays``.
-        """
-        convert_fn = "_{}_contract".format(backend)
 
+        If ``parse_constants=True``, perform a partial contraction that
+        prepares the constant tensors and operations with the right backend.
+        """
         # convert consts to correct type & find reduced contraction list
         if parse_constants:
             return backends.parse_constants(backend, arrays, self)
 
-        if not hasattr(self, convert_fn):
-            setattr(self, convert_fn, backends.build_expression(backend, arrays, self))
-
[…]
