### 2018-07-14 66d4b77
fix conversion when first arg is constant
