### 2018-07-20 cea9099
update ContractExpression __repr__
