### 2018-07-20 bd6ed18
first run at constants docs
