### 2018-08-03 96522ca
tweaks to constants incl docs
