+++ opt_einsum/contract.py
@@ -424,7 +424,7 @@ def contract(*operands, **kwargs):
     return _core_contract(operands, contraction_list, backend=backend, **einsum_kwargs)
 
 
-def _core_contract(operands, contraction_list, backend='numpy', parse_constants=False, **einsum_kwargs):
+def _core_contract(operands, contraction_list, backend='numpy', evaluate_constants=False, **einsum_kwargs):
     """Inner loop used to perform an actual contraction given the output
     from a ``contract_path(..., einsum_call=True)`` call.
     """
@@ -441,7 +441,7 @@ def _core_contract(operands, contraction_list, backend='numpy', parse_constants=
         inds, idx_rm, einsum_str, remaining, blas_flag = contraction
 
         # check if we are performing the pre-pass of an expression with constants
-        if parse_constants:
+        if evaluate_constants:
             # if so, keep contracting until we find any non-constant operators, then break out
             operands_copy = list(operands)
             if any(operands_copy.pop(x) is None for x in inds):
@@ -542,13 +542,13 @@ class ContractExpression:
         self._full_contraction_list = contraction_list
 
         self._constants_dict = constants_dict
-        self._parsed_constants = {}
+        self._evaluated_constants = {}
         self._backend_expressions = {}
 
-    def parse_constants(self, backend='numpy'):
+    def evaluate_constants(self, backend='numpy'):
         """Convert any constant operands to the correct backend form, and
         perform as many contractions as possible to create a new list of
-        operands, stored in ``self._parsed_constants[backend]``. This also
+        operands, stored in ``self._evaluated_constants[backend]``. This also
         makes sure ``self.contraction_list`` only contains the remaining,
         non-const operations.
         """
@@ -556,20 +556,20 @@ class ContractExpression:
         tmp_const_ops = [self._constants_dict.get(i, None) for i in range(self._full_num_args)]
 
         # get the new list of operands with constant operations performed, and remaining contractions
-        new_ops, new_contraction_list = self(*tmp_const_ops, backend=backend, parse_constants=True)
-        self._parsed_constants[backend] = new_ops
+        new_ops, new_contraction_list = self(*tmp_const_ops, backend=backend, evaluate_constants=True)
+        self._evaluated_constants[backend] = new_ops
         self.contraction_list = new_contraction_list
 
-    def _get_parsed_constants(self, backend):
+    def _get_evaluated_constants(self, backend):
         """Retrieve or generate the cached list of constant operators (mixed
         in with None representing non-consts) and the remaining contraction
         list.
         """
         try:
-            return self._parsed_constants[backend]
+            return self._evaluated_constants[backend]
         except KeyError:
-            self.parse_constants(backend)
-            return self._parsed_constants[backend]
+            self.evaluate_constants(backend)
+            return self._evaluated_constants[backend]
 
     def _get_backend_expression(self, arrays, backend):
         try:
@@ -579,26 +579,26 @@ class ContractExpression:
             self._backend_expressions[backend] = fn
             return fn
 
-    def _contract(self, arrays, out=None, backend='numpy', parse_constants=False):
+    def _contract(self, arrays, out=None, backend='numpy', evaluate_constants=False):
         """The normal, core contraction.
         """
-        contraction_list = self._full_contraction_list if parse_constants else self.contraction_list
+        contraction_list = self._full_contraction_list if evaluate_constants else self.contraction_list
 
         return _core_contract(list(arrays), contraction_list, out=out, backend=backend,
-                              parse_constants=parse_constants, **self.einsum_kwargs)
+                              evaluate_constants=evaluate_constants, **self.einsum_kwargs)
 
[…]
