### 2018-08-03 b17efdb
modularize backends, add pytorch, tensorflow eager support
