+++ opt_einsum/backends/dispatch.py
@@ -1,0 +1,101 @@
+"""
+Performing array conversions and operations with various backends.
+"""
+
+import importlib
+
+import numpy
+
+from . import cupy as _cupy
+from . import torch as _torch
+from . import theano as _theano
+from . import tensorflow as _tensorflow
+
+# known non top-level imports
+_aliases = {
+    'dask': 'dask.array',
+    'theano': 'theano.tensor',
+    'torch': 'opt_einsum.backends.torch',
+}
+
+
+def _import_func(func, backend):
+    """Try and import ``{backend}.{func}``, raise an error if library is
+    installed but can't find that particular function.
+    """
+    try:
+        lib = importlib.import_module(_aliases.get(backend, backend))
+        return getattr(lib, func)
+    except AttributeError:
+        raise AttributeError("{} doesn't seem to provide the function {}".format(backend, func))
+
+
+# manually cache functions as python2 doesn't support functools.lru_cache
+#     other libs will be added to this if needed, but pre-populate with numpy
+_cached_funcs = {
+    ('tensordot', 'numpy'): numpy.tensordot,
+    ('transpose', 'numpy'): numpy.transpose,
+    ('einsum', 'numpy'): numpy.einsum,
+}
+
+
+def get_func(func, backend='numpy'):
+    """Return ``{backend}.{func}``, e.g. ``numpy.einsum``, cache result.
+    """
+    try:
+        return _cached_funcs[func, backend]
+    except KeyError:
+        fn = _import_func(func, backend)
+        _cached_funcs[func, backend] = fn
+        return fn
+
+
+# mark libs with einsum, else try to use tensordot/tranpose as much as possible
+_has_einsum = {}
+
+
+def has_einsum(backend):
+    """Check if ``{backend}.einsum`` exists, cache result for performance.
+    """
+    try:
+        return _has_einsum[backend]
+    except KeyError:
+        try:
+            get_func('einsum', backend)
+            _has_einsum[backend] = True
+        except AttributeError:
+            _has_einsum[backend] = False
+
+        return _has_einsum[backend]
+
+
+# Dispatch to correct expression backend
+#    these are the backends which support explicit to-and-from numpy conversion
+CONVERT_BACKENDS = {
+    'tensorflow': _tensorflow.build_expression,
+    'theano': _theano.build_expression,
+    'cupy': _cupy.build_expression,
+    'torch': _torch.build_expression,
+}
+
+
+EVAL_CONSTS_BACKENDS = {
+    'tensorflow': _tensorflow.evaluate_constants,
+    'theano': _theano.evaluate_constants,
+    'cupy': _cupy.evaluate_constants,
+    'torch': _torch.evaluate_constants,
+}
+
+
+def build_expression(backend, arrays, expr):
+    """Build an expression, based on ``expr`` and initial arrays ``arrays``,
+    that evaluates using backend ``backend``.
+    """
+    return CONVERT_BACKENDS[backend](arrays, expr)
+
+
+def evaluate_constants(backend, arrays, expr):
+    """Convert constant arrays to the correct backend, and perform as much of
+    the contraction of ``expr`` with these as possible.
+    """
+    return EVAL_CONSTS_BACKENDS[backend](arrays, expr)
