### 2018-08-03 474bfe0
allow blas when both dimensions are 1
- previously it was assumed these were going to be broadcast
- also make a few more 'micro' optimizations, mostly changing appends to comprehensions
