### 2018-08-04 5ecd0b7
simplify core_contract constants break out
relies on the fact that ``inds`` are reverse sorted
