### 2018-08-04 1b7a095
py2 compat and update travis with pytorch
travis: quiet pip install


travis fixes

including python2 specific
fixes
