+++ opt_einsum/backends/cupy.py
@@ -1,3 +1,7 @@
+"""
+Required functions for optimized contractions of numpy arrays using cupy.
+"""
+
 from __future__ import absolute_import
 import numpy as np
 
