### 2018-08-06 9b04faa
CI: Switches to conda for environment testing
