### 2018-08-06 72d2d63
CI: No need for MKL during testing
