### 2018-08-06 c922a04
CI: Adds theano back into tests
