+++ devtools/conda-envs/full_test.yaml
@@ -10,6 +10,7 @@ dependencies:
   - tensorflow
   - dask
   - sparse
+  - theano
 
     # Testing
   - pytest
