### 2018-08-06 ac354a9
CI: Switches to conda for environment testing (#33)
* CI: Switches to conda for environment testing

* CI: No need for MKL during testing

* CI: Adds theano back into tests
