### 2018-08-06 f867ee3
travis: install pytorch-cpu
