### 2018-08-08 2dde998
update broadcast logic and allow blas on 'healed' indices
