+++ opt_einsum/tests/test_edge_cases.py
@@ -2,12 +2,10 @@
 Tets a series of opt_einsum contraction paths to ensure the results are the same for different paths
 """
 
-import sys
-
 import numpy as np
 import pytest
 
-from opt_einsum import contract, contract_path, helpers, contract_expression
+from opt_einsum import contract, contract_expression
 
 
 def test_contract_expression_checks():
@@ -114,3 +112,14 @@ def test_broadcasting_contraction4():
     opt = contract('obk,ijk->ioj', a, a, optimize=True)
 
     assert np.allclose(ein, opt)
+
+
+def test_can_blas_on_healed_broadcast_dimensions():
+
+    expr = contract_expression("ab,bc,bd->acd", (5, 4), (1, 5), (4, 20))
+    # first contraction involves broadcasting
+    assert expr.contraction_list[0][2] == 'bc,ab->bca'
+    assert expr.contraction_list[0][-1] is False
+    # but then is healed GEMM is usable
+    assert expr.contraction_list[1][2] == 'bca,bd->acd'
+    assert expr.contraction_list[1][-1] == 'GEMM'
