### 2018-08-09 c0541b0
travis: fix torch
don't use pytest importorskip
