### 2018-08-09 43e740f
for intermediaries use tensordot not einsum output order
minimises transpositions / maximises array contiguity
