### 2018-08-13 ac12a8a
tweaks for test coverage
