+++ opt_einsum/helpers.py
@@ -117,11 +117,7 @@ def find_contraction(positions, input_sets, output_set):
     """
 
     remaining = list(input_sets)
-    idx_contract = set()
-
-    for i in sorted(positions, reverse=True):
-        idx_contract |= remaining.pop(i)
-
+    idx_contract = set.union(*map(remaining.pop, sorted(positions, reverse=True)))
     idx_remain = output_set.union(*remaining)
 
     new_result = idx_remain & idx_contract
