### 2018-08-13 f8ab187
add 'auto' mode and make it the default
