### 2018-08-13 1cacec8
mention tensorflow and pytorch explicitly
