### 2018-08-13 e79f9ac
tweak to readability find_contraction
