### 2018-08-13 f381d16
Adds citation information to README and docs
