### 2018-08-14 3ea01c8
Updates release.md
