### 2018-08-15 676e505
Docs: Changes syntax theme to 'friendly'
