+++ docs/source/conf.py
@@ -78,9 +78,9 @@ author = 'Daniel G. A. Smith'
 # built documents.
 #
 # The short X.Y version.
-version = '1.0.1'
+version = opt_einsum.__version__
 # The full version, including alpha/beta/rc tags.
-release = '1.0.1'
+release = opt_einsum.__version__
 
 # The language for content autogenerated by Sphinx. Refer to documentation
 # for a list of supported languages.
@@ -123,7 +123,8 @@ exclude_patterns = ['_build', 'Thumbs.db', '.DS_Store', "units"]
 # show_authors = False
 
 # The name of the Pygments (syntax highlighting) style to use.
-pygments_style = 'sphinx'
+# ['default', 'emacs', 'friendly', 'colorful', 'autumn', 'murphy', 'manni', 'monokai', 'perldoc', 'pastie', 'borland', 'trac', 'native', 'fruity', 'bw', 'vim', 'vs', 'tango', 'rrt', 'xcode', 'igor', 'paraiso-light', 'paraiso-dark', 'lovelace', 'algol', 'algol_nu', 'arduino', 'rainbow_dash', 'abap']
+pygments_style = 'friendly'
 
 # A list of ignored prefixes for module index sorting.
 # modindex_common_prefix = []
