### 2018-08-15 e0219aa
Docs: Adds a changelog
