### 2018-08-15 6c0b1bb
Docs: Fixes oe.contract signature bug
