### 2018-08-15 8719d8a
Docs: ensures syntax highlighting in path finding
