### 2018-08-15 7ef12a5
Adds MANIFEST for versioneer deployment
