### 2018-08-17 20fa63e
Rename symbols to support torch.einsum
