+++ opt_einsum/backends/torch.py
@@ -5,7 +5,7 @@ Required functions for optimized contractions of numpy arrays using pytorch.
 from __future__ import absolute_import
 import numpy as np
 
-from ..parser import einsum_symbols_base
+from ..parser import einsum_symbols_base, get_symbol
 
 
 _TORCH_DEVICE = None
@@ -31,6 +31,12 @@ def transpose(a, axes):
 def einsum(equation, *operands):
     """Variadic version of torch.einsum to match numpy api.
     """
+    # rename symbols to support PyTorch 0.4.1 and earlier,
+    # which allow only symbols a-z.
+    symbols = sorted(set(equation) - set(',->'))
+    rename = {s: get_symbol(i) for i, s in enumerate(symbols)}
+    equation = ''.join(rename.get(s, s) for s in equation)
+
     torch, _ = _get_torch_and_device()
     return torch.einsum(equation, operands)
 
@@ -39,8 +45,6 @@ def tensordot(x, y, axes=2):
     """Simple translation of tensordot syntax to einsum.
     """
     # XXX: tensordot should be directly implemented in torch soon
-    torch, _ = _get_torch_and_device()
-
     xnd = x.ndimension()
     ynd = y.ndimension()
 
@@ -80,7 +84,7 @@ def tensordot(x, y, axes=2):
 
     # form full string and contract!
     einsum_str = "{},{}->{}".format(*map("".join, (x_ix, y_ix, out_ix)))
-    return torch.einsum(einsum_str, (x, y))
+    return einsum(einsum_str, x, y)
 
 
 def to_torch(array):
