### 2018-08-17 a24dfe0
Add test for upper-case characters
