### 2018-08-18 2656141
Move implementation to convert_to_valid_einsum_chars
