### 2018-08-19 dd6b34a
WIP Attempt to fix sharing
