+++ opt_einsum/__init__.py
@@ -2,6 +2,7 @@
 Main init function for opt_einsum.
 """
 
+from .backends.shared import shared_intermediates
 from .contract import contract, contract_path, contract_expression
 from .parser import get_symbol
 from . import paths

[…]
