### 2018-08-20 560726e
Cache tensor inputs to avoid id clash
