### 2018-08-20 589950b
Address review comments
