+++ opt_einsum/backends/shared.py
@@ -6,7 +6,7 @@ from ..parser import get_symbol
 from .dispatch import get_func
 
 _SHARING_STACK = []
-_CURRENT_BACKEND = []
+_CURRENT_BACKEND = None
 
 
 @contextlib.contextmanager
@@ -33,20 +33,26 @@ def shared_intermediates(cache=None):
     """
     if cache is None:
         cache = {}
-    _SHARING_STACK.append(cache)
-    yield cache
-    _SHARING_STACK.pop()
+    try:
+        _SHARING_STACK.append(cache)
+        yield cache
+    finally:
+        _SHARING_STACK.pop()
 
 
 @contextlib.contextmanager
 def handle_sharing(backend):
-    if _SHARING_STACK and not _CURRENT_BACKEND:
-        _CURRENT_BACKEND.append(backend)
-        yield 'opt_einsum.backends.shared'
-        _CURRENT_BACKEND.pop()
-    if backend == 'opt_einsum.backends.shared' and not _CURRENT_BACKEND:
+    global _CURRENT_BACKEND
+    if _SHARING_STACK and _CURRENT_BACKEND is None:
+        try:
+            _CURRENT_BACKEND = backend
+            yield __name__
+        finally:
+            _CURRENT_BACKEND = None
+    elif backend == __name__ and _CURRENT_BACKEND is None:
         raise ValueError('shared backend is available only via shared_intermediates')
-    yield backend
+    else:
+        yield backend
 
 
 def _alpha_canonicalize(equation):
@@ -61,56 +67,65 @@ def _alpha_canonicalize(equation):
     return ''.join(rename.get(x, x) for x in equation)
 
 
-def transpose(a, axes):
-    backend = _CURRENT_BACKEND[0]
+def _save_tensors(*tensors):
+    """Save tensors in the cache to prevent their ids from being recycled.
+    This is needed to prevent false cache lookups.
+    """
     cache = _SHARING_STACK[-1]
-    cache['tensor', id(a)] = a
+    for tensor in tensors:
+        cache['tensor', id(tensor)] = tensor
 
+
+def transpose(a, axes):
+    _save_tensors(a)
+
+    # hash by axes
     axes = tuple(axes)
-    key = 'transpose', backend, id(a), axes
+    key = 'transpose', _CURRENT_BACKEND, id(a), axes
+
+    cache = _SHARING_STACK[-1]
     if key in cache:
         return cache[key]
 
-    result = get_func('transpose', backend)(a, axes)
+    result = get_func('transpose', _CURRENT_BACKEND)(a, axes)
     cache[key] = result
     return result
 
 
 def tensordot(x, y, axes=2):
-    backend = _CURRENT_BACKEND[0]
-    cache = _SHARING_STACK[-1]
-    cache['tensor', id(x)] = x
-    cache['tensor', id(y)] = y
+    _save_tensors(x, y)
 
+    # hash based on the (axes_x,axes_y) form of axes
     if isinstance(axes, numbers.Number):
         axes = list(range(len(x.shape)))[len(x.shape) - axes:], list(range(len(y.shape)))[:axes]
     axes = tuple(axes[0]), tuple(axes[1])
-    key = 'tensordot', backend, id(x), id(y), axes
+    key = 'tensordot', _CURRENT_BACKEND, id(x), id(y), axes
+
+    cache = _SHARING_STACK[-1]
     if key in cache:
         return cache[key]
 
-    result = get_func('tensordot', backend)(x, y, axes)
+    result = get_func('tensordot', _CURRENT_BACKEND)(x, y, axes)
     cache[key] = result
     return result
 
 
 def einsum(equation, *operands):
-    backend = _CURRENT_BACKEND[0]
-    cache = _SHARING_STACK[-1]
-    for d in operands:
[…]
