### 2018-08-21 5ff111f
Refactor to use function wrappers
