### 2018-08-21 72b9ef8
Memoize conversions to backend
