+++ opt_einsum/tests/test_sharing.py
@@ -1,7 +1,8 @@
 import numpy as np
 import pytest
 
-from opt_einsum import contract, contract_path, get_symbol, shared_intermediates
+from opt_einsum import (contract_expression, contract_path, get_symbol,
+                        helpers, shared_intermediates)
 
 try:
     import torch
@@ -9,39 +10,45 @@ try:
 except ImportError:
     torch_if_found = pytest.param('torch', marks=[pytest.mark.skip(reason="PyTorch not installed.")])
 
+backends = ['numpy', torch_if_found]
 
-@pytest.mark.parametrize('backend', ['numpy', torch_if_found])
+
+def contract(eq, *ops, **kwargs):
+    expr = contract_expression(eq, *(op.shape for op in ops))
+    return expr(*ops, **kwargs)
+
+
+@pytest.mark.parametrize('backend', backends)
 def test_sharing_value(backend):
-    w = np.random.normal(size=(2, 3, 4))
-    x = np.random.normal(size=(3, 4, 5))
-    y = np.random.normal(size=(4, 5, 6))
-    z = np.random.normal(size=(5, 6, 7))
-    expr = 'abc,bcd,cde,def->af'
+    eq = 'abc,bcd,cde,def->af'
+    views = helpers.build_views(eq)
+    shapes = [v.shape for v in views]
+    expr = contract_expression(eq, *shapes)
 
-    expected = contract(expr, w, x, y, z, backend=backend)
+    expected = expr(*views, backend=backend)
     with shared_intermediates():
-        actual = contract(expr, w, x, y, z, backend=backend)
+        actual = expr(*views, backend=backend)
 
     assert (actual == expected).all()
 
 
-@pytest.mark.parametrize('backend', ['numpy', torch_if_found])
+@pytest.mark.parametrize('backend', backends)
 def test_complete_sharing(backend):
-    x = np.random.normal(size=(5, 4))
-    y = np.random.normal(size=(4, 3))
-    z = np.random.normal(size=(3, 2))
+    eq = 'ab,bc,cd->'
+    views = helpers.build_views(eq)
+    expr = contract_expression(eq, *(v.shape for v in views))
 
     print('-' * 40)
     print('Without sharing:')
     with shared_intermediates() as cache:
-        contract('ab,bc,cd->', x, y, z, backend=backend)
+        expr(*views, backend=backend)
         expected = len(cache)
 
     print('-' * 40)
     print('With sharing:')
     with shared_intermediates() as cache:
-        contract('ab,bc,cd->', x, y, z, backend=backend)
-        contract('ab,bc,cd->', x, y, z, backend=backend)
+        expr(*views, backend=backend)
+        expr(*views, backend=backend)
         actual = len(cache)
 
     print('-' * 40)
@@ -50,28 +57,28 @@ def test_complete_sharing(backend):
     assert actual == expected
 
 
-@pytest.mark.parametrize('backend', ['numpy', torch_if_found])
+@pytest.mark.parametrize('backend', backends)
 def test_partial_sharing(backend):
-    x = np.random.normal(size=(5, 4))
-    y = np.random.normal(size=(4, 3))
-    z1 = np.random.normal(size=(3, 2))
-    z2 = np.random.normal(size=(3, 2))
+    eq = 'ab,bc,cd->'
+    x, y, z1 = helpers.build_views(eq)
+    z2 = 2.0 * z1 - 1.0
+    expr = contract_expression(eq, x.shape, y.shape, z1.shape)
 
     print('-' * 40)
     print('Without sharing:')
     num_exprs_nosharing = 0
     with shared_intermediates() as cache:
-        contract('ab,bc,cd->', x, y, z1, backend=backend)
[…]
