### 2018-08-21 d208e3d
Make tests more idiomatic
