### 2018-08-21 472dbff
Add count_cached_ops method for profiling
