### 2018-08-21 4fb5ae5
Add test of commutativity
