### 2018-08-21 e771973
Add test of reused cache
