### 2018-08-21 632c266
Support full equation syntax in sharing
