### 2018-08-21 b3f2394
Move decorators to contract.py
