### 2018-08-21 a6983c6
Add test for nested sharing
