### 2018-08-21 911eca4
Move alpha_canonicalize to parser.py
