### 2018-08-21 86d4550
Add test of sharing with constants
