+++ opt_einsum/tests/test_sharing.py
@@ -242,6 +242,27 @@ def compute_cost(cache):
     return sum(1 for key in cache.keys() if key[0] in ('einsum', 'tensordot'))
 
 
+@pytest.mark.parametrize('backend', backends)
+def test_sharing_with_constants(backend):
+    inputs = 'ij,jk,kl'
+    outputs = 'ijkl'
+    equations = ['{}->{}'.format(inputs, output) for output in outputs]
+    shapes = (2, 3), (3, 4), (4, 5)
+    constants = {0, 2}
+    ops = [np.random.rand(*shp) if i in constants else shp for i, shp in enumerate(shapes)]
+    var = np.random.rand(*shapes[1])
+
+    expected = [contract_expression(eq, *shapes)(ops[0], var, ops[2])
+                for eq in equations]
+
+    with shared_intermediates():
+        actual = [contract_expression(eq, *ops, constants=constants)(var)
+                  for eq in equations]
+
+    for dim, expected_dim, actual_dim in zip(outputs, expected, actual):
+        assert np.allclose(expected_dim, actual_dim), 'error at {}'.format(dim)
+
+
 @pytest.mark.parametrize('size', [3, 4, 5])
 @pytest.mark.parametrize('backend', backends)
 def test_chain(size, backend):
