### 2018-08-21 0139c75
Add bullet point to readme
