### 2018-08-22 e43d68b
Address review comments
