+++ docs/source/sharing_intermediates.rst
@@ -8,29 +8,34 @@ For example, suppose we want to compute marginals at each point in a factor chai
 
 .. code:: python
 
-    >>> inputs = 'ab,bc,cd,de,ef'
-    >>> factors = [np.random.rand(4, 4) for _ in range(5)]
-    >>> marginals = {output: contract('{}->{}'.format(inputs, output), *factors)
-    >>>              for output in 'abcdef'}
+    inputs = 'ab,bc,cd,de,ef'
+    factors = [np.random.rand(1000, 1000) for _ in range(5)]
+
+    %%timeit
+    marginals = {output: contract('{}->{}'.format(inputs, output), *factors)
+                 for output in 'abcdef'}
+    1 loop, best of 3: 5.82 s per loop
 
 To share this computation, we can perform all contractions in a shared context
 
 .. code:: python
 
-    >>> with shared_intermediates():
-    >>>     marginals = {output: contract('{}->{}'.format(inputs, output), *factors)
-    >>>                  for output in 'abcdef'}
+    %%timeit
+    with shared_intermediates():
+        marginals = {output: contract('{}->{}'.format(inputs, output), *factors)
+                     for output in 'abcdef'}
+    1 loop, best of 3: 1.55 s per loop
 
 If it is difficult to fit your code into a context, you can instead save the sharing cache for later reuse.
 
 .. code:: python
 
-    >>> with shared_intermediates() as cache:  # create a cache
-    >>>     pass
-    >>> marginals = {}
-    >>> for output in 'abcdef':
-    >>>     with shared_intermediates(cache):  # reuse a common cache
-    >>>         marginals[output] = contract('{}->{}'.format(inputs, output), *factors)
-    >>> del cache  # garbage collect intermediates
+    with shared_intermediates() as cache:  # create a cache
+        pass
+    marginals = {}
+    for output in 'abcdef':
+        with shared_intermediates(cache):  # reuse a common cache
+            marginals[output] = contract('{}->{}'.format(inputs, output), *factors)
+    del cache  # garbage collect intermediates
 
 Note that sharing contexts can be nested, so it is safe to to use :func:`~opt_einsum.shared_intermediates` in library code without leaking intermediates into user caches.
