### 2018-08-23 cb509c5
Avoid unicode in README.md; avoid failing in setup.py
