### 2018-08-23 44ada3e
Avoid unicode in README.md; avoid failing in setup.py
