### 2018-08-23 cec1030
Make get_symbol(large_number) work in Python2
