### 2018-08-23 176b405
Enable fixed tests
