### 2018-08-23 20ce7ef
Simplify shim
