### 2018-08-23 eb74ad3
Enable more tests
