### 2018-08-23 9ad4a75
Make get_symbol(large_number) work in Python2 (#48)
* Make get_symbol(large_number) work in Python2

* Enable fixed tests

* Simplify shim

* Enable more tests

* Avoid .format() in einsum_cache_wrap
