### 2018-08-24 20ba7e8
switch _SHARING_STACK to dict
