+++ opt_einsum/sharing.py
@@ -4,10 +4,37 @@ import contextlib
 import functools
 import numbers
 from collections import Counter
+import threading
 
 from .parser import alpha_canonicalize, parse_einsum_input
 
-_SHARING_STACK = []
+
+_SHARING_STACK = {}
+
+
+def currently_sharing():
+    """Check if we are currently sharing a cache -- thread specific.
+    """
+    return threading.get_ident() in _SHARING_STACK
+
+
+def get_sharing_cache():
+    """Return the most recent sharing cache -- thread specific.
+    """
+    return _SHARING_STACK[threading.get_ident()][-1]
+
+
+def _add_sharing_cache(cache):
+    tid = threading.get_ident()
+    _SHARING_STACK.setdefault(tid, [])
+    _SHARING_STACK[tid].append(cache)
+
+
+def _remove_sharing_cache(cache):
+    tid = threading.get_ident()
+    _SHARING_STACK[tid].remove(cache)
+    if not _SHARING_STACK[tid]:
+        del _SHARING_STACK[tid]
 
 
 @contextlib.contextmanager
@@ -34,11 +61,11 @@ def shared_intermediates(cache=None):
     """
     if cache is None:
         cache = {}
+    _add_sharing_cache(cache)
     try:
-        _SHARING_STACK.append(cache)
         yield cache
     finally:
-        _SHARING_STACK.pop()
+        _remove_sharing_cache(cache)
 
 
 def count_cached_ops(cache):
@@ -52,7 +79,7 @@ def _save_tensors(*tensors):
     """Save tensors in the cache to prevent their ids from being recycled.
     This is needed to prevent false cache lookups.
     """
-    cache = _SHARING_STACK[-1]
+    cache = get_sharing_cache()
     for tensor in tensors:
         cache['tensor', id(tensor)] = tensor
 
@@ -62,7 +89,7 @@ def _memoize(key, fn, *args, **kwargs):
     Results will be stored in the innermost ``cache`` yielded by
     :func:`shared_intermediates`.
     """
-    cache = _SHARING_STACK[-1]
+    cache = get_sharing_cache()
     if key in cache:
         return cache[key]
     result = fn(*args, **kwargs)
@@ -77,7 +104,7 @@ def transpose_cache_wrap(transpose):
 
     @functools.wraps(transpose)
     def cached_transpose(a, axes, backend='numpy'):
-        if not _SHARING_STACK:
+        if not currently_sharing():
             return transpose(a, axes, backend=backend)
 
         # hash by axes
@@ -96,7 +123,7 @@ def tensordot_cache_wrap(tensordot):
 
     @functools.wraps(tensordot)
     def cached_tensordot(x, y, axes=2, backend='numpy'):
-        if not _SHARING_STACK:
+        if not currently_sharing():
             return tensordot(x, y, axes, backend=backend)
 
         # hash based on the (axes_x,axes_y) form of axes
@@ -117,7 +144,7 @@ def einsum_cache_wrap(einsum):
 
     @functools.wraps(einsum)
     def cached_einsum(*args, **kwargs):
-        if not _SHARING_STACK:
+        if not currently_sharing():
             return einsum(*args, **kwargs)
 
         # hash modulo commutativity by computing a canonical ordering and names
@@ -143,7 +170,7 @@ def to_backend_cache_wrap(to_backend):
 
     @functools.wraps(to_backend)
     def cached_to_backend(array):
-        if not _SHARING_STACK:
+        if not currently_sharing():
             return to_backend(array)
 
         # hash by id
