### 2018-08-24 8fe7027
address comments, py2 compat
