+++ opt_einsum/sharing.py
@@ -3,36 +3,41 @@
 import contextlib
 import functools
 import numbers
-from collections import Counter
+from collections import Counter, defaultdict
 import threading
 
 from .parser import alpha_canonicalize, parse_einsum_input
 
 
-_SHARING_STACK = {}
+_SHARING_STACK = defaultdict(list)
+
+
+try:
+    get_thread_id = threading.get_ident
+except AttributeError:
+    def get_thread_id():
+        return threading.current_thread().ident
 
 
 def currently_sharing():
     """Check if we are currently sharing a cache -- thread specific.
     """
-    return threading.get_ident() in _SHARING_STACK
+    return get_thread_id() in _SHARING_STACK
 
 
 def get_sharing_cache():
     """Return the most recent sharing cache -- thread specific.
     """
-    return _SHARING_STACK[threading.get_ident()][-1]
+    return _SHARING_STACK[get_thread_id()][-1]
 
 
 def _add_sharing_cache(cache):
-    tid = threading.get_ident()
-    _SHARING_STACK.setdefault(tid, [])
-    _SHARING_STACK[tid].append(cache)
+    _SHARING_STACK[get_thread_id()].append(cache)
 
 
-def _remove_sharing_cache(cache):
-    tid = threading.get_ident()
-    _SHARING_STACK[tid].remove(cache)
+def _remove_sharing_cache():
+    tid = get_thread_id()
+    _SHARING_STACK[tid].pop()
     if not _SHARING_STACK[tid]:
         del _SHARING_STACK[tid]
 
@@ -65,7 +70,7 @@ def shared_intermediates(cache=None):
     try:
         yield cache
     finally:
-        _remove_sharing_cache(cache)
+        _remove_sharing_cache()
 
 
 def count_cached_ops(cache):
