### 2018-08-25 89997c3
make shared_intermediates thread-safe (#53)
* switch _SHARING_STACK to dict

* address comments, py2 compat

* test multi-threaded sharing
