### 2018-08-25 51bd2c0
Canonicalization pass
