+++ opt_einsum/backends/cupy.py
@@ -3,8 +3,12 @@ Required functions for optimized contractions of numpy arrays using cupy.
 """
 
 from __future__ import absolute_import
+
 import numpy as np
+
 from ..sharing import to_backend_cache_wrap
+
+__all__ = ["to_cupy", "build_expression", "evaluate_constants"]
 
 
 @to_backend_cache_wrap
@@ -31,4 +35,4 @@ def evaluate_constants(const_arrays, expr):  # pragma: no cover
     """Convert constant arguments to cupy arrays, and perform any possible
     constant contractions.
     """
-    return expr(*[to_cupy(x) for x in const_arrays], backend='cupy', evaluate_constants=True)
+    return expr(* [to_cupy(x) for x in const_arrays], backend='cupy', evaluate_constants=True)
