### 2018-08-25 a4a9cf0
Inline doc comments
