+++ opt_einsum/sharing.py
@@ -165,11 +165,15 @@ def einsum_cache_wrap(einsum):
         equation = args[0]
         inputs, output, operands = parse_einsum_input(args)
         inputs = inputs.split(',')
+
         _save_tensors(*operands)
+
+        # Build canonical key
         canonical = sorted(zip(inputs, map(id, operands)), key=lambda x: x[1])
         canonical_ids = tuple(id_ for _, id_ in canonical)
         canonical_inputs = ','.join(input_ for input_, _ in canonical)
         canonical_equation = alpha_canonicalize(canonical_inputs + "->" + output)
+
         key = 'einsum', backend, canonical_equation, canonical_ids
         return _memoize(key, einsum, equation, *operands, backend=backend)
 
