### 2018-08-25 85ed2f8
General sphinx doc update pass
