+++ docs/source/ex_dask_reusing_intermediaries.rst
@@ -4,10 +4,10 @@ Reusing Intermediaries with Dask
 `Dask <https://dask.pydata.org/>`_ provides a computational framework where
 arrays and the computations on them are built up into a 'task graph' before
 computation. Since :mod:`opt_einsum` is compatible with ``dask`` arrays this
-means that mutiple contractions can be built into the same task graph, which
+means that multiple contractions can be built into the same task graph, which
 then automatically reuses any shared arrays and contractions.
 
-For example imagine the two expressions:
+For example, imagine the two expressions:
 
 .. code-block:: python
 
@@ -31,7 +31,7 @@ First, let's set up some ``numpy`` arrays:
     >>> np_ops1 = [np_arrays[s] for s in terms1]
     >>> np_ops2 = [np_arrays[s] for s in terms2]
 
-Normally we would just compute these separately:
+Typically we would compute these expressions separately:
 
 .. code-block:: python
 
@@ -41,7 +41,7 @@ Normally we would just compute these separately:
     >>> oe.contract(contraction2, *np_ops2)
     array(-75.55902751)
 
-But if we use dask arrays we can combine the two operations, so let's set those
+However, if we use dask arrays we can combine the two operations, so let's set those
 up:
 
 .. code-block:: python
@@ -72,17 +72,17 @@ Note ``chunks`` is a required argument relating to how the arrays are stored (se
     >>> dy
     Delayed('list-3af82335-b75e-47d6-b800-68490fc865fd')
 
-As suggested by the name ``Delayed``, we just have a placeholder for the result
-so far. When we want to actually *perform* the computation we can call:
+As suggested by the name ``Delayed``, we have a placeholder for the result
+so far. When we want to *perform* the computation we can call:
 
 .. code-block:: python
 
     >>> dy.compute()
     [114.78314052155015, -75.55902750513113]
 
-Which matches the numpy result. The computation can even be handled by various
+The above matches the canonical numpy result. The computation can even be handled by various
 schedulers - see `scheduling <http://dask.pydata.org/en/latest/scheduling.html>`_.
-Finally, to check we really are reusing intermediaries we can view the task
+Finally, to check we are reusing intermediaries, we can view the task
 graph generated for the computation:
 
 .. code-block:: python
@@ -90,3 +90,6 @@ graph generated for the computation:
     >>> dy.visualize(optimize_graph=True)
 
 .. image:: ex_dask_reuse_graph.png
+
+.. note::
+    For sharing intermediates with other backends see `Sharing Intermediates <sharing_intermediates.html>`_. Dask graphs are particularly useful for reusing intermediates beyond just contractions and can allow additional parallelization.

[…]
