### 2018-08-26 6de2017
Intermediate sharing result rework
