### 2018-08-27 3d72bba
v2.2 doc cleanup (#54)
* Canonicalization pass

* Inline doc comments

* General sphinx doc update pass

* Intermediate sharing result rework
