+++ opt_einsum/tests/test_contract.py
@@ -124,7 +124,7 @@ def test_compare_greek(string):
     ein = contract(string, *views, optimize=False, use_blas=False)
 
     # convert to greek
-    string = ''.join(compat.get_chr(ord(c) + 848) if c not in ',->.' else c for c in string)
+    string = ''.join(compat.get_char(ord(c) + 848) if c not in ',->.' else c for c in string)
 
     opt = contract(string, *views, optimize='greedy', use_blas=False)
     assert np.allclose(ein, opt)
@@ -152,7 +152,7 @@ def test_compare_blas_greek(string):
     ein = contract(string, *views, optimize=False)
 
     # convert to greek
-    string = ''.join(compat.get_chr(ord(c) + 848) if c not in ',->.' else c for c in string)
+    string = ''.join(compat.get_char(ord(c) + 848) if c not in ',->.' else c for c in string)
 
     opt = contract(string, *views, optimize='greedy')
     assert np.allclose(ein, opt)
@@ -163,7 +163,7 @@ def test_compare_blas_greek(string):
 
 def test_some_non_alphabet_maintains_order():
     # 'c beta a' should automatically go to -> 'a c beta'
-    string = 'c' + compat.get_chr(ord('b') + 848) + 'a'
+    string = 'c' + compat.get_char(ord('b') + 848) + 'a'
     # but beta will be temporarily replaced with 'b' for which 'cba->abc'
     # so check manual output kicks in:
     x = np.random.rand(2, 3, 4)
