### 2018-08-28 574e7af
allow sharing for tensorflow and theano as numpy backends
