+++ opt_einsum/backends/theano.py
@@ -6,9 +6,12 @@ from __future__ import absolute_import
 
 import numpy as np
 
+from ..sharing import to_backend_cache_wrap
+
 __all__ = ["to_theano", "build_expression", "evaluate_constants"]
 
 
+@to_backend_cache_wrap(constants=True)
 def to_theano(array, constant=False):
     """Convert a numpy array to ``theano.tensor.TensorType`` instance.
     """
