### 2018-08-29 1117f09
allow sharing for tensorflow and theano as numpy backends (#58)
* allow sharing for tensorflow and theano as numpy backends

* fix: constants wasn't propagating into cache call
