### 2018-09-11 18cb60a
LGTM: Adds configuration badge
