+++ .lgtm.yml
@@ -1,0 +1,7 @@
+# Configure LGTM for this package
+
+path_classifiers:
+  library:
+    - versioneer.py  # Set Versioneer.py to an external "library" (3rd party code)
+  generated:
+    - opt_einsum/_version.py
