### 2018-09-11 a9e5fb5
LGTM: Minor LGTM tweaks
