+++ .lgtm.yml
@@ -3,5 +3,6 @@
 path_classifiers:
   library:
     - versioneer.py  # Set Versioneer.py to an external "library" (3rd party code)
+    - setup.py
   generated:
     - opt_einsum/_version.py
