### 2018-09-12 1db5a0a
Minor tweaks (#59)
* LGTM: Adds configuration badge

* LGTM: Minor LGTM tweaks
