### 2018-09-30 d382e3d
Add prototype cheap optimizer
