+++ opt_einsum/paths.py
@@ -1,11 +1,18 @@
 """
 Contains the path technology behind opt_einsum in addition to several path helpers
 """
+from __future__ import absolute_import, division, print_function
+
+import heapq
 import itertools
+from collections import defaultdict
+
+import numpy as np
 
 from . import helpers
 
-__all__ = ["optimal", "greedy"] 
+__all__ = ["optimal", "greedy", "cheap"]
+
 
 def optimal(input_sets, output_set, idx_dict, memory_limit):
     """
@@ -272,3 +279,214 @@ def greedy(input_sets, output_set, idx_dict, memory_limit):
         path_cost += best[0][1]
 
     return path
+
+
+def ssa_to_linear(ssa_path):
+    """
+    Convert a path with static single assignment ids to a path with recycled
+    linear ids.
+    """
+    ids = np.arange(sum(map(len, ssa_path)), dtype=np.int64)
+    path = []
+    for i, ssa_ids in enumerate(ssa_path):
+        path.append(tuple(ids[ssa_id] for ssa_id in ssa_ids))
+        for ssa_id in ssa_ids:
+            ids[ssa_id:] -= 1
+    return path
+
+
+def linear_to_ssa(path):
+    """
+    Convert a path with recycled linear ids to a path with static single
+    assignment ids.
+    """
+    num_inputs = sum(map(len, path)) - len(path) + 1
+    linear_to_ssa = list(range(num_inputs))
+    new_ids = itertools.count(num_inputs)
+    ssa_path = []
+    for ids in path:
+        ssa_path.append(tuple(linear_to_ssa[id_] for id_ in ids))
+        for id_ in sorted(ids, reverse=True):
+            del linear_to_ssa[id_]
+        linear_to_ssa.append(next(new_ids))
+    return ssa_path
+
+
+def _get_candidate(output, sizes, remaining, footprints, dim_ref_counts, k1, k2):
+    either = k1 | k2
+    two = k1 & k2
+    one = either - two
+    k12 = (either & output) | (two & dim_ref_counts[3]) | (one & dim_ref_counts[2])
+    footprint12 = helpers.compute_size_by_dict(k12, sizes)
+    cost = footprint12 - footprints[k1] - footprints[k2]
+    id1 = remaining[k1]
+    id2 = remaining[k2]
+    cost = cost, min(id1, id2), max(id1, id2)  # break ties to ensure determinism
+    return cost, k1, k2, k12
+
+
+def _push_candidate(output, sizes, remaining, footprints, dim_ref_counts, k1, k2s, queue):
+    if not k2s:
+        return
+    candidate = min(_get_candidate(output, sizes, remaining, footprints, dim_ref_counts, k1, k2)
+                    for k2 in k2s)
+    heapq.heappush(queue, candidate)
+
+
+def _update_ref_counts(dim_to_keys, dim_ref_counts, dims):
+    for dim in dims:
+        count = len(dim_to_keys[dim])
+        if count <= 1:
+            dim_ref_counts[2].discard(dim)
+            dim_ref_counts[3].discard(dim)
+        elif count == 2:
+            dim_ref_counts[2].add(dim)
+            dim_ref_counts[3].discard(dim)
+        else:
+            dim_ref_counts[2].add(dim)
+            dim_ref_counts[3].add(dim)
+
+
+def _ssa_optimize(inputs, output, sizes):
+    """
+    This has an interface similar to :func:`optimize` but produces a path with
+    static single assignment ids rather than recycled linear ids.
+    SSA ids are cheaper to work with and easier to reason about.
+    """
+    # A dim common to all terms might as well be an output dim.
+    inputs = list(map(frozenset, inputs))
+    output = frozenset(output) | frozenset.intersection(*inputs)
+
[…]
