### 2018-09-30 8a8bf82
Fix docstring
