### 2018-09-30 6bb352e
Simplify objective function; improve sharing
