### 2018-10-01 8fec17a
path finding changes
fix cupy a bit
