### 2018-10-01 f8bdece
PathInfo: inherit from object and use join for repr
