+++ opt_einsum/contract.py
@@ -15,7 +15,7 @@ from . import sharing
 __all__ = ["contract_path", "contract", "format_const_einsum_str", "ContractExpression", "shape_only", "shape_only"]
 
 
-class PathInfo:
+class PathInfo(object):
 
     def __init__(self, contraction_list, input_subscripts, output_subscript,
                  indices, scale_list, naive_cost, opt_cost, size_list):
@@ -41,24 +41,26 @@ class PathInfo:
         overall_contraction = self.input_subscripts + "->" + self.output_subscript
         header = ("scaling", "BLAS", "current", "remaining")
 
-        path_print = "  Complete contraction:  {}\n".format(overall_contraction)
-        path_print += "         Naive scaling:  {}\n".format(len(self.indices))
-        path_print += "     Optimized scaling:  {}\n".format(max(self.scale_list))
-        path_print += "      Naive FLOP count:  {:.3e}\n".format(naive_cost)
-        path_print += "  Optimized FLOP count:  {:.3e}\n".format(opt_cost)
-        path_print += "   Theoretical speedup:  {:3.3f}\n".format(speedup)
-        path_print += "  Largest intermediate:  {:.3e} elements\n".format(largest_intermediate)
-        path_print += "-" * 80 + "\n"
-        path_print += "{:>6} {:>11} {:>22} {:>37}\n".format(*header)
-        path_print += "-" * 80
+        path_print = [
+            "  Complete contraction:  {}\n".format(overall_contraction),
+            "         Naive scaling:  {}\n".format(len(self.indices)),
+            "     Optimized scaling:  {}\n".format(max(self.scale_list)),
+            "      Naive FLOP count:  {:.3e}\n".format(naive_cost),
+            "  Optimized FLOP count:  {:.3e}\n".format(opt_cost),
+            "   Theoretical speedup:  {:3.3f}\n".format(speedup),
+            "  Largest intermediate:  {:.3e} elements\n".format(largest_intermediate),
+            "-" * 80 + "\n",
+            "{:>6} {:>11} {:>22} {:>37}\n".format(*header),
+            "-" * 80
+        ]
 
         for n, contraction in enumerate(self.contraction_list):
             inds, idx_rm, einsum_str, remaining, do_blas = contraction
             remaining_str = ",".join(remaining) + "->" + self.output_subscript
             path_run = (self.scale_list[n], do_blas, einsum_str, remaining_str)
-            path_print += "\n{:>4} {:>14} {:>22} {:>37}".format(*path_run)
+            path_print.append("\n{:>4} {:>14} {:>22} {:>37}".format(*path_run))
 
-        return path_print
+        return "".join(path_print)
 
 
 def _choose_memory_arg(memory_limit, size_list):
