+++ opt_einsum/paths.py
@@ -312,23 +312,24 @@ def linear_to_ssa(path):
     return ssa_path
 
 
-def _get_candidate(output, sizes, remaining, footprints, dim_ref_counts, k1, k2):
+def _get_candidate(output, sizes, remaining, dim_ref_counts, k1, k2):
     either = k1 | k2
     two = k1 & k2
     one = either - two
     k12 = (either & output) | (two & dim_ref_counts[3]) | (one & dim_ref_counts[2])
-    footprint12 = helpers.compute_size_by_dict(k12, sizes)
-    cost = footprint12 - footprints[k1] - footprints[k2]
+    cost = helpers.compute_size_by_dict(k12, sizes)
     id1 = remaining[k1]
     id2 = remaining[k2]
-    cost = cost, min(id1, id2), max(id1, id2)  # break ties to ensure determinism
+    if id1 > id2:
+        k1, id1, k2, id2 = k2, id2, k1, id1
+    cost = cost, id1, id2  # break ties to ensure determinism
     return cost, k1, k2, k12
 
 
-def _push_candidate(output, sizes, remaining, footprints, dim_ref_counts, k1, k2s, queue):
+def _push_candidate(output, sizes, remaining, dim_ref_counts, k1, k2s, queue):
     if not k2s:
         return
-    candidate = min(_get_candidate(output, sizes, remaining, footprints, dim_ref_counts, k1, k2)
+    candidate = min(_get_candidate(output, sizes, remaining, dim_ref_counts, k1, k2)
                     for k2 in k2s)
     heapq.heappush(queue, candidate)
 
@@ -353,7 +354,14 @@ def _ssa_optimize(inputs, output, sizes):
     static single assignment ids rather than recycled linear ids.
     SSA ids are cheaper to work with and easier to reason about.
     """
-    # A dim common to all terms might as well be an output dim.
+    if len(inputs) == 1:
+        # Perform a single contraction to match output shape.
+        return [(0,)]
+
+    # A dim that is common to all tensors might as well be an output dim, since it
+    # cannot be contracted until the final step. This avoids an expensive all-pairs
+    # comparison to search for possible contractions at each step, leading to speedup
+    # in many practical problems where all tensors share a common batch dimension.
     inputs = list(map(frozenset, inputs))
     output = frozenset(output) | frozenset.intersection(*inputs)
 
@@ -367,10 +375,6 @@ def _ssa_optimize(inputs, output, sizes):
             remaining[key] = next(ssa_ids)
         else:
             remaining[key] = ssa_id
-
-    # Compute footprints of each tensor.
-    footprints = {key: helpers.compute_size_by_dict(key, sizes)
-                  for key in remaining}
 
     # Keep track of possible contraction dims.
     dim_to_keys = defaultdict(set)
@@ -388,10 +392,10 @@ def _ssa_optimize(inputs, output, sizes):
     # Find initial candidate contractions.
     queue = []
     for dim, keys in dim_to_keys.items():
-        keys = list(keys)
+        keys = sorted(keys, key=remaining.__getitem__)
         for i, k1 in enumerate(keys):
-            k2s = keys[:i]
-            _push_candidate(output, sizes, remaining, footprints, dim_ref_counts, k1, k2s, queue)
+            k2s = keys[1 + i:]
+            _push_candidate(output, sizes, remaining, dim_ref_counts, k1, k2s, queue)
 
     # Greedily contract pairs of tensors.
     while queue:
@@ -405,11 +409,10 @@ def _ssa_optimize(inputs, output, sizes):
[…]
