### 2018-10-01 9d540b4
update docstrings and var names
