### 2018-10-01 ddd47b7
update string formatting
