### 2018-10-01 809a3b1
test rand_equation
