### 2018-10-01 f297d5f
set seed for rand_equation test
