### 2018-10-02 cd68094
Add test for linear-ssa id conversions
