### 2018-10-02 ccbdf6b
various changes to path machinery (#61)
* path finding changes

* fix cupy a bit

* PathInfo: inherit from object and use join for repr

* update docstrings and var names

* update string formatting

* test rand_equation

* set seed for rand_equation test
