### 2018-10-03 8c63e4b
Rename cheap to eager
