+++ docs/source/api.rst
@@ -11,6 +11,6 @@ Function Reference
     opt_einsum.contract.ContractExpression
     opt_einsum.paths.optimal
     opt_einsum.paths.greedy
-    opt_einsum.paths.cheap
+    opt_einsum.paths.eager
     opt_einsum.parser.get_symbol
     opt_einsum.sharing.shared_intermediates
