### 2018-10-03 524ea12
Include memory pressure in objective function
