### 2018-10-03 cb5119b
Simplify objective function and optimize
