+++ opt_einsum/paths.py
@@ -284,7 +284,10 @@ def greedy(input_sets, output_set, idx_dict, memory_limit):
 def ssa_to_linear(ssa_path):
     """
     Convert a path with static single assignment ids to a path with recycled
-    linear ids.
+    linear ids. For example::
+
+        >>> ssa_to_linear([(0, 3), (1, 4), (2, 5)])
+        [(0, 3), (1, 2), (0, 1)]
     """
     ids = np.arange(sum(map(len, ssa_path)), dtype=np.int64)
     path = []
@@ -298,7 +301,10 @@ def ssa_to_linear(ssa_path):
 def linear_to_ssa(path):
     """
     Convert a path with recycled linear ids to a path with static single
-    assignment ids.
+    assignment ids. For example::
+
+        >>> linear_to_ssa([(0, 3), (1, 2), (0, 1)])
+        [(0, 3), (1, 4), (2, 5)]
     """
     num_inputs = sum(map(len, path)) - len(path) + 1
     linear_to_ssa = list(range(num_inputs))
