### 2018-10-04 2fd09a4
Misc cleanup
