### 2018-10-04 9219d67
Fix docstrings
