+++ opt_einsum/paths.py
@@ -318,12 +318,12 @@ def linear_to_ssa(path):
     return ssa_path
 
 
-def _get_candidate(output, sizes, remaining, costs, dim_ref_counts, k1, k2):
+def _get_candidate(output, sizes, remaining, footprints, dim_ref_counts, k1, k2):
     either = k1 | k2
     two = k1 & k2
     one = either - two
     k12 = (either & output) | (two & dim_ref_counts[3]) | (one & dim_ref_counts[2])
-    cost = costs[k1] + costs[k2] + helpers.compute_size_by_dict(k12, sizes)
+    cost = helpers.compute_size_by_dict(k12, sizes) - footprints[k1] - footprints[k2]
     id1 = remaining[k1]
     id2 = remaining[k2]
     if id1 > id2:
@@ -332,8 +332,8 @@ def _get_candidate(output, sizes, remaining, costs, dim_ref_counts, k1, k2):
     return cost, k1, k2, k12
 
 
-def _push_candidate(output, sizes, remaining, costs, dim_ref_counts, k1, k2s, queue):
-    candidate = min(_get_candidate(output, sizes, remaining, costs, dim_ref_counts, k1, k2)
+def _push_candidate(output, sizes, remaining, footprints, dim_ref_counts, k1, k2s, queue):
+    candidate = min(_get_candidate(output, sizes, remaining, footprints, dim_ref_counts, k1, k2)
                     for k2 in k2s)
     heapq.heappush(queue, candidate)
 
@@ -394,7 +394,7 @@ def _ssa_optimize(inputs, output, sizes):
         for count in [2, 3]}
 
     # Compute separable part of the objective function for contractions.
-    costs = {key: -0.5 * helpers.compute_size_by_dict(key, sizes) for key in remaining}
+    footprints = {key: helpers.compute_size_by_dict(key, sizes) for key in remaining}
 
     # Find initial candidate contractions.
     queue = []
@@ -402,7 +402,7 @@ def _ssa_optimize(inputs, output, sizes):
         keys = sorted(keys, key=remaining.__getitem__)
         for i, k1 in enumerate(keys[:-1]):
             k2s = keys[1 + i:]
-            _push_candidate(output, sizes, remaining, costs, dim_ref_counts, k1, k2s, queue)
+            _push_candidate(output, sizes, remaining, footprints, dim_ref_counts, k1, k2s, queue)
 
     # Greedily contract pairs of tensors.
     while queue:
@@ -424,14 +424,14 @@ def _ssa_optimize(inputs, output, sizes):
                 dim_to_keys[dim].add(k12)
         remaining[k12] = next(ssa_ids)
         _update_ref_counts(dim_to_keys, dim_ref_counts, k1 | k2 - output)
-        costs[k12] = -0.5 * helpers.compute_size_by_dict(k12, sizes)
+        footprints[k12] = helpers.compute_size_by_dict(k12, sizes)
 
         # Find new candidate contractions.
         k1 = k12
         k2s = set(k2 for dim in k1 for k2 in dim_to_keys[dim])
         k2s.discard(k1)
         if k2s:
-            _push_candidate(output, sizes, remaining, costs, dim_ref_counts, k1, k2s, queue)
+            _push_candidate(output, sizes, remaining, footprints, dim_ref_counts, k1, k2s, queue)
 
     # Greedily compute pairwise outer products.
     queue = [(helpers.compute_size_by_dict(key, sizes), ssa_id, key)
