### 2018-10-05 f4ef3ed
Fix python2 unicode bug in PathInfo
