### 2018-10-05 bd9a61c
Add prototype cheap optimizer (#60)
* Add prototype cheap optimizer

* Fix docstring

* Disable failing test

* Simplify objective function; improve sharing

* Address review comments

* Fix test typo

* Add test for linear-ssa id conversions

* Rename cheap to eager

* Include memory pressure in objective function

* Simplify objective function and optimize

* Add docs

* Misc cleanup

* Fix docstrings

* Address review comments

* Fix python2 unicode bug in PathInfo
