### 2018-10-16 02d15b7
torch: use tensordot if available (#66)
