### 2018-10-18 1bf74be
Support more symbols for subscripts (#67)
* Add more symbol support based on hash

* Refactor the converting to subscripts routine a little bit

* Update tests
