### 2018-10-18 ea4c2f0
Remove redundant debug code
