### 2018-10-19 33a09d6
Fix incorrect output index order due to hash
* Sort the symbol map to avoid order perturbation

* Add tests on the output index order when using list subscripts
