### 2018-10-19 855efd1
Add primitive document for user-custom subscript object
* Add a new chapter "input_format" and put it in the toc

* Fix unused import defaultdict

* Add comments for the user-custom subscript code

* Add a new test regarding output subscripts
