### 2018-10-20 7daf290
Support more symbols for subscripts (#67) (#68)
* Support more symbols for subscripts (#67)

* Add more symbol support based on hash

* Refactor the converting to subscripts routine a little bit

* Update tests

* Remove redundant debug code

* Fix incorrect output index order due to hash

* Sort the symbol map to avoid order perturbation

* Add tests on the output index order when using list subscripts

* Add primitive document for user-custom subscript object

* Add a new chapter "input_format" and put it in the toc

* Fix unused import defaultdict

* Add comments for the user-custom subscri
[…]
