### 2018-10-20 aa47e5b
input fixes and docs
