### 2018-10-20 8c885b6
fix for python2 syntax compatibility
