### 2018-10-21 4476a8e
Tweak input docs to be more explicit
