### 2018-10-21 9a03438
initial recursive optimal
