### 2018-10-21 dbf284b
input fixes and docs (#72)
* input fixes and docs

* fix for python2 syntax compatibility

* Tweak input docs to be more explicit
