### 2018-10-23 9ccbc3c
add recursive greedy implementation
