+++ opt_einsum/tests/test_contract.py
@@ -97,8 +97,12 @@ tests = [
 ]
 
 
+all_optimizers = ['greedy', 'optimal', 'eager', 'roptimal',
+                  'rgreedy', 'rgreedy1', 'rgreedy2']
+
+
 @pytest.mark.parametrize("string", tests)
-@pytest.mark.parametrize("optimize", ['greedy', 'optimal', 'eager'])
+@pytest.mark.parametrize("optimize", all_optimizers)
 def test_compare(optimize, string):
     views = helpers.build_views(string)
 
@@ -115,7 +119,7 @@ def test_drop_in_replacement(string):
 
 
 @pytest.mark.parametrize("string", tests)
-@pytest.mark.parametrize("optimize", ['greedy', 'optimal', 'eager'])
+@pytest.mark.parametrize("optimize", all_optimizers)
 def test_compare_greek(optimize, string):
     views = helpers.build_views(string)
 
@@ -129,7 +133,7 @@ def test_compare_greek(optimize, string):
 
 
 @pytest.mark.parametrize("string", tests)
-@pytest.mark.parametrize("optimize", ['greedy', 'optimal', 'eager'])
+@pytest.mark.parametrize("optimize", all_optimizers)
 def test_compare_blas(optimize, string):
     views = helpers.build_views(string)
 
@@ -139,7 +143,7 @@ def test_compare_blas(optimize, string):
 
 
 @pytest.mark.parametrize("string", tests)
-@pytest.mark.parametrize("optimize", ['greedy', 'optimal', 'eager'])
+@pytest.mark.parametrize("optimize", all_optimizers)
 def test_compare_blas_greek(optimize, string):
     views = helpers.build_views(string)
 
@@ -170,7 +174,7 @@ def test_printing():
 
 
 @pytest.mark.parametrize("string", tests)
-@pytest.mark.parametrize("optimize", ['greedy', 'optimal', 'eager'])
+@pytest.mark.parametrize("optimize", all_optimizers)
 @pytest.mark.parametrize("use_blas", [False, True])
 @pytest.mark.parametrize("out_spec", [False, True])
 def test_contract_expressions(string, optimize, use_blas, out_spec):
