### 2018-10-30 ce3163e
tidy up and replace optimal/greedy
update docs
