+++ opt_einsum/paths.py
@@ -11,17 +11,92 @@ import numpy as np
 
 from . import helpers
 
-__all__ = ["optimal", "greedy", "eager"]
+__all__ = ["optimal", "branch", "greedy", "eager"]
 
 
-inf = float('inf')
-
-
-def optimal(input_sets, output_set, idx_dict, memory_limit):
+def ssa_to_linear(ssa_path):
     """
-    Computes all possible pair contractions, sieves the results based
-    on ``memory_limit`` and returns the lowest cost path. This algorithm
-    scales factorial with respect to the elements in the list ``input_sets``.
+    Convert a path with static single assignment ids to a path with recycled
+    linear ids. For example::
+
+        >>> ssa_to_linear([(0, 3), (2, 4), (1, 5)])
+        [(0, 3), (1, 2), (0, 1)]
+    """
+    ids = np.arange(1 + max(map(max, ssa_path)), dtype=np.int32)
+    path = []
+    for ssa_ids in ssa_path:
+        path.append(tuple(int(ids[ssa_id]) for ssa_id in ssa_ids))
+        for ssa_id in ssa_ids:
+            ids[ssa_id:] -= 1
+    return path
+
+
+def linear_to_ssa(path):
+    """
+    Convert a path with recycled linear ids to a path with static single
+    assignment ids. For example::
+
+        >>> linear_to_ssa([(0, 3), (1, 2), (0, 1)])
+        [(0, 3), (2, 4), (1, 5)]
+    """
+    num_inputs = sum(map(len, path)) - len(path) + 1
+    linear_to_ssa = list(range(num_inputs))
+    new_ids = itertools.count(num_inputs)
+    ssa_path = []
+    for ids in path:
+        ssa_path.append(tuple(linear_to_ssa[id_] for id_ in ids))
+        for id_ in sorted(ids, reverse=True):
+            del linear_to_ssa[id_]
+        linear_to_ssa.append(next(new_ids))
+    return ssa_path
+
+
+def _calc_k12_flops(inputs, output, remaining, i, j, size_dict):
+    """Calculate the resulting indices and flops for a potential pairwise
+    contraction - used in the recursive (optimal/branch) algorithms.
+
+    Parameters
+    ----------
+    inputs : tuple[frozenset[str]]
+        The indices of each tensor in this contraction, note this includes
+        tensors unavaiable to contract as static single assignment is used ->
+        contracted tensors are not removed from the list.
+    output : frozenset[str]
+        The set of output indices for the whole contraction.
+    remaining : frozenset[int]
+        The set of indices (corresponding to ``inputs``) of tensors still
+        available to contract.
+    i : int
+        Index of potential tensor to contract.
+    j : int
+        Index of potential tensor to contract.
+    size_dict dict[str, int]
+        Size mapping of all the indices.
+
+    Returns
+    -------
+    k12 : frozenset
+        The resulting indices of the potential tensor.
+    cost : int
+        Estimated flop count of operation.
+    """
+    k1, k2 = inputs[i], inputs[j]
+    either = k1 | k2
+    shared = k1 & k2
+    keep = frozenset.union(output, *map(inputs.__getitem__, remaining - {i, j}))
+
+    k12 = either & keep
+    cost = helpers.flop_count(either, shared - keep, 2, size_dict)
+
+    return k12, cost
+
+
+def optimal(inputs, output, size_dict, memory_limit=None):
+    """
+    Computes all possible pair contractions in a depth-first recursive manner,
+    sieving results based on ``memory_limit`` and the best path found so far.
+    Returns the lowest cost path. This algorithm scales factoriallly with
+    respect to the elements in the list ``input_sets``.
 
[…]
