### 2018-10-30 d1f0bd5
implement memory_limit for optimal & branch
