### 2018-11-03 443b352
docs and tweaks
