+++ docs/source/reusing_paths.rst
@@ -96,7 +96,8 @@ two contractions to compute the output.
 .. note::
 
     The constant part of an expression is lazily generated upon the first call
-    (specific to each backend), though it can also be explicitly built by calling:meth:`~opt_einsum.contract.ContractExpression.evaluate_constants`.
+    (specific to each backend), though it can also be explicitly built by calling
+    :meth:`~opt_einsum.contract.ContractExpression.evaluate_constants`.
 
 We can confirm the advantage of using expressions and constants by timing the
 following scenarios, first setting
