### 2018-11-21 d4a0592
Allow overriding of transpose method
